+++ rer/cookieconsent/upgrades/upgrades.py
@@ -2,12 +2,17 @@
 from plone import api
 
 
-DEFAULT_PROFILE = 'profile-rer.cookieconsent:default'
+DEFAULT_PROFILE = "profile-rer.cookieconsent:default"
 
 
 def to_2000(context):
-    'Import new bundle and remove old resources'
-    setup_tool = api.portal.get_tool('portal_setup')
-    setup_tool.runImportStepFromProfile(DEFAULT_PROFILE, 'plone.app.registry')
-    setup_tool.runAllImportStepsFromProfile(
-        'profile-rer.cookieconsent:to_2000')
+    "Import new bundle and remove old resources"
+    setup_tool = api.portal.get_tool("portal_setup")
+    setup_tool.runImportStepFromProfile(DEFAULT_PROFILE, "plone.app.registry")
+    setup_tool.runAllImportStepsFromProfile("profile-rer.cookieconsent:to_2000")
+
+
+def to_2001(context):
+    "Fix bundle registration"
+    setup_tool = api.portal.get_tool("portal_setup")
+    setup_tool.runImportStepFromProfile(DEFAULT_PROFILE, "plone.app.registry")